# swagger_ui/openapi.yml
openapi: 3.0.2
info:
  title: Question Augmentation - A Swagger UI/web service
  description: 'A Swagger UI/web service interface for the Question augmentation service'
  version: '0.0.1'
  contact:
    email: whomever@renci.org
  termsOfService: http://robokop.renci.org:7055/tos?service_long=Question+augmentation+service&provider_long=the+Translator+Consortium
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
tags:
  - name: Interfaces
paths:
  /node_expand:
    post:
      tags:
        - Interfaces
      summary: 'Get similar nodes related to a Biolink model formatted question.'
      description: 'Returns a list of similar nodes related to the question passed in.'
      requestBody:
        content:
          application/json:
              schema:
                "$ref": "#/definitions/Question"
        description: A question graph
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/definitions/Question"
          description: 'Example: Returns an array of expanded nodes and edges'

  /edge_expand:
    post:
      tags:
        - Interfaces
      summary: 'Get similar edges related to a Biolink model formatted question.'
      description: 'Returns a list of similar edges related to the question passed in.'
      parameters:
        - in: query
          name: depth
          required: true
          default: 1
          example: 1
      requestBody:
        content:
          application/json:
              schema:
                "$ref": "#/definitions/Question"
        description: A question graph
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/definitions/Question"
          description: 'Example: Returns an array of expanded edges and nodes'

definitions:
  QEdge:
    description: Minimal edge specification
    properties:
      id:
        description: Internal ID of this edge
        type: string
      source_id:
        description: Internal ID of source node of this edge
        type: string
      target_id:
        description: Internal ID of target node of this edge
        type: string
      type:
        description: Optional relationship type of this edge
        type: string
    required:
    - id
    - source_id
    - target_id
    type: object
  QGraph:
    description: Graph representing the minimal question specification
    properties:
      edges:
        items:
          "$ref": "#/definitions/QEdge"
        type: array
      nodes:
        items:
          "$ref": "#/definitions/QNode"
        type: array
    type: object
  QNode:
    description: Minimal node specification
    properties:
      curie:
        description: Optional curie of this node
        type: string
      id:
        description: Internal ID of this node
        type: string
      set:
        description: Flag indicating whether node should be collapsed in answers
        type: boolean
      type:
        description: Optional biolink-model type of this node
        type: string
    required:
    - id
  Question:
    description: Question
    example:
      machine_question:
        edges:
        - id: edge_0
          source_id: node_0
          target_id: node_1
          type: increases_transport_of
        - id: edge_1
          source_id: node_1
          target_id: node_2
          type: contributes_to
        nodes:
        - id: node_0
          type: gene
        - id: node_1
          type: chemical_substance
        - id: node_2
          type: disease
      name: question name
      natural_question: what question is this?
      notes: Some notes about the question
    id: Question
    properties:
      machine_question:
        "$ref": "#/definitions/QGraph"
      name:
        description: human-readable question name
        type: string
      natural_question:
        description: natural language English question
        type: string
      notes:
        description: notes about question
        type: string